backend/requirements.txt
Flask
Flask-Cors
yfinance
requests
google-generativeai
numpy
pandas
scikit-learn
tensorflow
ta
gunicorn
